M1 -- corrige numpy.float64 vazando pro JSON e QLIKE=inf mascarando o vencedor
diebold_mariano::dm_stat vazava numpy.float64 (np.sqrt(n) nao envolvido
em float()), invisivel nos 23 testes existentes (pytest.approx nao liga
pro tipo) mas quebrava orjson.dumps na primeira vez que o valor saiu do
laboratorio de teste pra virar relatorio real.

Achado mais serio: qlike_mean pooled (media ignorando so NaN) virava
`inf` em TODAS as 15 combinacoes, baseline e candidatos igualmente --
uma unica barra com forecast quase-zero (9 a 1882 por combinacao, medido)
basta pra contaminar a media inteira, porque nanmean ignora NaN mas nao
ignora inf. `qlike_mean < baseline.qlike_mean` era sempre falso por
construcao, tornando `stopping_criterion_1_triggered=True` um artefato
da metrica quebrada, nao um resultado real -- Diebold-Mariano (que
diferencia ponto-a-ponto e cancela os pontos degenerados quando os dois
lados sao inf no mesmo bar) ja mostrava o oposto.

Corrigido: qlike_mean (pooled e por fold) agora exclui observacoes nao-
finitas explicitamente, com n_inf_qlike reportado para auditoria -- nao
e suavizar a metrica, e a diferenca entre ela comparar alguma coisa ou
nunca comparar nada.

Resultado real (relatorio anexo): Parkinson e Garman-Klass batem o
baseline ATRWilder em QLIKE nas 15/15 combinacoes (fold_win_rate ~1.0,
DM p~=0), tanto nos 20 folds do BTCUSDT (inclui 2019-2021) quanto nos 12
dos 4 alts (so 2021-2026) -- sinal nao e artefato da janela mais longa
do BTC. RealizedVol perde em 15/15. stopping_criterion_1_triggered=False
-- linha de busca de volatilidade continua aberta (Sec6.5).

+++ src/validation/volatility_walkforward.py
@@ -235,6 +235,6 @@ def diebold_mariano(loss_candidate: FloatArray, loss_baseline: FloatArray) -> Di
     std_d = float(np.std(valid, ddof=1))
     if std_d == 0.0:
         return DieboldMarianoResult(float("nan"), float("nan"), mean_d, n)
-    dm_stat = mean_d / (std_d / np.sqrt(n))
+    dm_stat = float(mean_d / (std_d / np.sqrt(n)))
     p_value = float(2.0 * (1.0 - norm.cdf(abs(dm_stat))))
     return DieboldMarianoResult(dm_stat, p_value, mean_d, n)
